requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anthropic==0.34.0
anyio==3.7.1
async-timeout==4.0.3
attrs==23.2.0
cachetools==5.5.0
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
exceptiongroup==1.2.0
filelock==3.13.1
frozenlist==1.4.1
fsspec==2023.12.1
google-ai-generativelanguage==0.6.6
google-api-core==2.19.2
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-generativeai==0.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.0
grpcio-status==1.62.3
h11==0.14.0
httpcore==1.0.2
httplib2==0.22.0
httpx==0.27.0
huggingface-hub==0.19.4
idna==3.6
Jinja2==3.1.2
jiter==0.5.0
jsonpath-python==1.0.6
MarkupSafe==2.1.3
mistralai==1.0.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.2
openai==1.3.7
packaging==23.2
proto-plus==1.24.0
protobuf==4.25.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.8.2
pydantic_core==2.20.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
rsa==4.9
safetensors==0.4.1
scipy==1.11.4
sentencepiece==0.1.99
six==1.16.0
sniffio==1.3.0
tenacity==8.2.3
termcolor==2.4.0
text-generation==0.6.1
tiktoken==0.5.2
tokenizers==0.15.0
tqdm==4.66.1
transformers==4.38.2
typing-inspect==0.9.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
yarl==1.9.4
